+++ core/src/com/pottda/game/model/ActorFactory.java
@@ -1,7 +1,6 @@
 package com.pottda.game.model;
 
 import com.pottda.game.controller.AbstractController;
-import com.pottda.game.view.Sprites;
 
 import javax.vecmath.Vector2f;
