+++ core/src/com/pottda/game/model/ModelActor.java
@@ -1,7 +1,5 @@
 package com.pottda.game.model;
 
-import com.pottda.game.view.Sprites;
-
 import javax.vecmath.Vector2f;
