+++ core/src/com/pottda/game/model/builders/AbstractModelBuilder.java
@@ -2,9 +2,8 @@ package com.pottda.game.model.builders;
 
 import com.pottda.game.model.ModelActor;
 import com.pottda.game.model.NewModelListener;
-import com.pottda.game.model.PhysicsActor;
 import com.pottda.game.model.PhysicsActorFactory;
-import com.pottda.game.view.Sprites;
+import com.pottda.game.model.Sprites;
 
 import javax.vecmath.Vector2f;
 import java.util.ArrayList;
